refactor: extract ConnectRepo.vue
